requirements.txt
beautifulsoup4~=4.13.4
jaconv~=0.4.0
regex==2023.12.25
sudachipy==0.6.10
tqdm==4.66.3
dragonmapper~=0.3.0
PyYAML~=6.0.2
sudachidict_full
jamdict~=0.1a11.post2
jamdict-data
bs4~=0.0.2
lxml~=6.0.0